auto_completion added

+++ srcs/line_edition/vi_mode/vi_setmode.c
@@ -6,7 +6,7 @@
 /*   By: roliveir <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/10 16:34:56 by roliveir          #+#    #+#             */
-/*   Updated: 2019/05/03 11:20:34 by roliveir         ###   ########.fr       */
+/*   Updated: 2019/05/08 15:45:46 by oboutrol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int					vi_new_mode(t_emode mode)
 	i = 0;
 	while (i < MODE)
 	{
-		if (i == mode)
+		if (i == (int)mode)
 			g_env.mode->mode[i] = 1;
 		else
 			g_env.mode->mode[i] = 0;
